+++ app/utils/dbConn.js
@@ -1,0 +1,7 @@
+let { Pool } = require("pg");
+
+let env = require("../../env.json");
+
+let pool = new Pool(env);
+
+module.exports = pool;
